Remove service-auth plugin and integrate authentication in server.js
The service-auth plugin was removed, and its functionality was directly implemented in server.js. This simplifies the codebase by reducing indirection and centralizing the authentication logic. Corresponding tests were also deleted as they became redundant.
