+++ src/server.js
@@ -1,11 +1,11 @@
 import Hapi from '@hapi/hapi'
+import Basic from '@hapi/basic'
 import Inert from '@hapi/inert'
 import Vision from '@hapi/vision'
 import HapiSwagger from 'hapi-swagger'
 
 import { config } from './config.js'
 import { router } from './plugins/router.js'
-import { serviceAuth } from './plugins/service-auth.js'
 import { requestLogger } from './common/helpers/logging/request-logger.js'
 import { mongoDb } from './common/helpers/mongodb.js'
 import { failAction } from './common/helpers/fail-action.js'
@@ -52,13 +52,33 @@ async function createServer() {
     documentationPath: '/documentation'
   }
 
+  // Register Basic auth and configure strategy
+  await server.register(Basic)
+
+  const serviceCredentials = config.get('serviceCredentials')
+
+  server.auth.strategy('service-token', 'basic', {
+    validate: async (_request, username, password) => {
+      if (!serviceCredentials) {
+        return { isValid: false, credentials: { username } }
+      }
+
+      const matchingCredential = serviceCredentials.find(
+        (cred) => cred.username === username && cred.password === password
+      )
+
+      return { isValid: !!matchingCredential, credentials: { username } }
+    }
+  })
+
+  server.auth.default('service-token')
+
   // Hapi Plugins:
   // requestLogger  - automatically logs incoming requests
   // requestTracing - trace header logging and propagation
   // secureContext  - loads CA certificates from environment config
   // pulse          - provides shutdown handlers
   // mongoDb        - sets up mongo connection pool and attaches to `server` and `request` objects
-  // serviceAuth    - service-to-service authentication
   // router         - routes used in the app
   await server.register([
     Inert,
@@ -67,7 +87,6 @@ async function createServer() {
       plugin: HapiSwagger,
       options: swaggerOptions
     },
-    serviceAuth,
     requestLogger,
     requestTracing,
     secureContext,
